I think I have this right now.  At least it's no longer throwing warnings on my dev box.  Other than on in the lexer

+++ src/libparser/core.c
@@ -414,7 +414,7 @@ void free_list(object_type *obj) {
 	}
 	
 	free(obj);
-	printf("freed:%li\n", obj);
+	printf("freed:%ju\n", (uintmax_t)obj);
 	obj=next;
     }
 }
